Weapon functionality/Shotgun alteration
The weapons are now added to the player and enabled when the player swaps to them.
Configured firing animations for both weapons.
Decreased the shotgun firing speed to better fit the firing animation which is more accurate to real world values.
Changed the line trace channel back to Pawn until I fix the problem with a custom channel not reacting properly.