Enemy shooting setup
Added a weapon to the gas mask enemy class.
Added the weapon BP to the gas mask enemy BP.
The weapon currently fires whenever possible until it is depleted.
Note: The weapon currently functions exactly the same as the player's but it will be changed in the future (different damage target/stats).