Objective/Pickup changes
The objective panel is now hidden when the player does not have an objective.
Objectives can now be set to seek the player or be placed in the world to give the player the objective upon contact.
Both pickups are now triggered on overlap rather than hit so that the player is not momentarily impeded by the pickup.