Objective/Laser sight/Enemy changes
Changed the blueprints for the different objectives so that the setup is done there and they can be reused.
Added a name component to the parent enemy class but it is not currently being used.
Changed the location of the laser sight to the character mesh so that it tracks better