Melee
Added a melee ability to the player which deals damage around him in a small radius. 
Created a timer for this ability which is triggered upon use. The player cannot fire their weapon or melee again until this duration is over.
Imported a melee animation and created a montage for it.
Added logic to play the montage in the player class.