Player animations/Laser sight
Imported firing and reloading animations.
Created montages from those animations.
Added the montage slots to the player class so that they can be played when needed.
Added a function to the parent weapon class to detect when the weapon can be reloaded.
Added logic to play the montages when firing and reloading and configured them to play at different rates based on the current weapon's stats.
Added a timer and two functions in order to disable the laser sight when the player uses the melee ability and when they reload.
*Notes: Reloading currently occurs when weapon swapping so that will have to be taken into consideration for the laser sight however this might change in the future*