AI speed and improvements/Animation tweaks
Created a member variable in the gas mask enemy class for their speed.
The enemy's speed is now slower before they encounter the player for the first time.
Walking animations for the player and the enemy have been sped up to better reflect the movement.
The enemy will now attempt to reload their weapon after the player leaves their sight if needed.