Objective refactor/Temporary door
The player now has an objective at all times and it is updated with the values of the objective on the objective giver on contact.
Objectives no longer need to be placed on the world and are instead created in the objective giver blueprint.
Created a temporary door which blocks the player from progressing until the current objective is complete.