Enemy ragdoll death
Added a function to the parent enemy class to handle them taking damage.
This function takes optional parameters for direction of the hit and force of the hit.
When the enemy dies and goes into a ragdoll state it applies a force to the ragdoll based on the force and direction of the last hit it received.
Configured the force and direction for the shotgun and assault rifle.