Grenade explosion/Ragdoll force change
Added logic to the grenade that allows it to explode after a short duration or upon making contact with an enemy.
The grenade explosion will deal damage to any enemy in the radius.
The ragdoll direction calculations have been redone as the previous formula was not applying the correct force.