Grenade SFX and VFX/Health orb fix
Added two sounds to the grenade, one to be played when it comes into contact with a surface and one when it explodes.
Added a particle to be played when the grenade explodes.
Configured all of the above SFX and VFX.
Fixed the health globe pickup as it was being destroyed whenever an actor came into contact with it.