Enemy AI reload/Grenade
Changed the enemy reload behavior tree task to only reload if the enemy's gun is at less than half capacity.
Added an input binding for throwing the grenade.
Added a cooldown to the grenade.
The grenade is now spawned in the player class.
Imported a grenade throwing animation.
Created a montage for throwing the grenade.
Created timers for the grenade cooldown and for when the grenade object should be spawned.
Added a function to determine if the player can perform actions and removed a blueprint function that did something similar.
Disabled reload, melee and weapon swap while throwing the grenade.