Melee and grenade fix
The melee damage and grenade damage was previously done using sphere overlap actor and filtering for enemies. This had the unintended behavior of being able to hit enemies through walls. Both damage types are now changed to radial damage as it allows walls to block the damage. Applying radial damage is done in C++ but receiving it is currently implemented in blueprint.