Explosion Ragdoll
The GetHit function in the parent enemy class can now be called in blueprint.
In the enemy blueprint, the event ReceiveRadialDamage determines if there should be a force exerted in the enemy or if the enemy should be stunned. As a result, the explosion now adds a force to the enemy's ragdoll on kill.