Damage types/Radial damage
Added two new damage types, explosive and melee.
Added an override for TakeDamage in the parent enemy class which handles when the enemy gets hit by radial damage (grenade and melee).
Added logic to determine the damage type and to call separate functions depending on those types. *Note: the logic in GetHit could be done in the TakeDamage function*