Miniboss level changes
Changed the layout of the miniboss level to make more sense with the camera angle.
Added a miniboss door blueprint which teleports the player to the other side of the door and does not allow the player to leave. The door becomes inactive once the player has killed the corresponding miniboss.
Configured the red miniboss door material.
Added logic to the miniboss which teleports the player back to the starting room on death.