Minibosses
Created child classes for each miniboss type so that it will be easier to distinguish their behavior.
Changed the miniboss door to automatically set its color based on the miniboss behind it.
Changed the door materials to reflect this behavior.
Changed the miniboss minimap markers to automatically change color based on the color of the miniboss.
Remade all miniboss blueprints as their specific miniboss class rather than as children of the parent miniboss class.