Rocket Launcher/Player changes/General fixes
Created a rocket launcher weapon for the red miniboss to use. 
Created a rocket that the rocket launcher can fire.
Configured the animations and sounds for the rocket launcher and rocket.
Added the rocket launcher to the miniboss and allowed it to fire (currently in tick for testing).
Added two functions to the parent player which allow them to take damage and changed existing enemy damage sources to call those functions.
Changed the miniboss door to only trigger when interacting with the player.
Added the player to the grenade and melee ignore lists as the player can now receive radial damage and self damage is not a desired mechanic.