Red Miniboss AI/Damage type
Added random movement to the red miniboss before firing at the player.
Added a node to set the miniboss invulnerable.
Changed the damage type of the rocket to a new damage type "DamageType_Enemy". This is to prevent the enemy from doing self damage when a damage source is only intended to damage the player.