Red miniboss AI/Damage shield
Added new behavior tree tasks which set the miniboss immune to damage and back to being able to receive damage.
Imported a texture for the damage shield.
Created a particle system for the damage shield and attached it to the miniboss.
The damage shield becomes active when the miniboss is invulnerable and the color is set based on the color of the miniboss.