Behavior tree tasks/Falling rocket
Created behavior tree tasks that set and remove the focus of the miniboss in relation to the player and one to orient the miniboss toward the camera.
Created a falling rocket blueprint which is derived from the rocket class. The rocket will fall from a location and have a range visualization displayed on the floor so that the player can find a safe space to stand.