Miniboss level/Shield generator prep/Player damage changes
Changed the miniboss level so that all boss doors are in the starting room. This allows the player more freedom in choosing a battle.
Created a new class for non-enemy actors who can take damage and derived a shield generator class from that one. The shield generators are to be used in a miniboss encounter.
Changed the shotgun and assault rifle in order to be able to apply damage to any damageable actors.