Yellow miniboss/Shield generator
Added a damage shield to the yellow miniboss and configured it.
Added functions to enable/disable the shield on the shield generator.
Added a miniboss variable to the shield generator so that they can be connected to a specific miniboss.
Added a boolean to the damageable actor class which controls whether or not the actor can currently be damaged.
Added a damage shield to the shield generator blueprint and configured it to take the color of the miniboss that it can give shields to.