Yellow miniboss shield mechanic
Configured the miniboss to have an array of shield generators.
Added three shield generators to the miniboss arena and added them to the miniboss' array.
Added a value for a health threshold in the miniboss' AI controller which always triggers the next living shield generator to lose its shield.
When a shield generator loses its shield it gives it to the miniboss. This allows the shield generator to be damaged. Once the shield generator is destroyed, the boss becomes damageable again.