Yellow miniboss weapon
Added a weapon to the miniboss (same as the gas mask enemy).
Added functions to reload and fire the weapon in the miniboss class.
Configured the weapon in the miniboss blueprint
Changed the scale of the gas amsk enemy weapon in the blueprint rather than in the gas mask enemy blueprint.