Yellow miniboss AI/Weapon changes
Added a boolean in the blackboard to detect if the miniboss is dead and set it in the AI controller.
Added a boolean for the rotation direction when the miniboss sweeps while firing.
Changed the logic of two tasks in order to randomize the direction that the boss will sweep and to call the appropriate functions based on that.
Changed the miniboss weapon firing rate so that the boss is just about emptying their clip each time they sweep.
Changed a delay in the behavior tree to before the miniboss selects their starting rotation and not after.