Green miniboss weapon
Created a weapon for the green miniboss which is based off of the player's shotgun.
Created a blueprint from that weapon class and configured it.
Added the weapon to the miniboss' blueprint and configured it along with the weapon muzzle scene component.
Added functions for the miniboss which allow it to fire and reload. *Note: for testing purposes, the weapon is currently set to fire in the tick*