Green miniboss mechanic
Added logic to the miniboss AI controller so that it spawns weaker copies of itself in a radius around itself every time they lose 25% health.
Added logic to destroy any living clones when the miniboss dies.
Added a toggle to the miniboss class which can set whether or not the miniboss' death should trigger a teleportation of the player back to the starting zone.
Configured the spawned minibosses to not teleport the player on death.