Green miniboss clone
Created a new class for the miniboss clone which is derived from the miniboss class.
Set them to be weaker than the miniboss and removed their ability to teleport the player on death.
Created an AI controller, blackboard and behavior tree for the clones.
Changed the miniboss AI controller to instead spawn AI from class rather than an actor so that the clones can use the proper behavior tree and controller.
Configured the clone blueprint to match the miniboss.