Green miniboss/clone AI
Added a task to fire the shotgun.
Added a task to move to a random point near the player.
Added a task that stops the enemy's movement when they are in range of the player.
Added logic so that the enemies can only fire if they are within the range of their weapon.
Added booleans to each blackboard which track whether or not the enemy can see the player and whether or not they are in range of the player.
Configured the behavior trees.
Fixed a bug with the AI controllers that made it so that the enemy was sometimes losing sight of the player incorrectly.