Miniboss blue weapon/prep
Created and configured a class for the miniboss weapon (sniper rifle).
Created a blueprint from the class and attached it to the miniboss.
Added animations for the sniper rifle firing and reloading actions.
Configured the damage shield, weapon and reload animation in the miniboss blueprint.
Change the AI controller so that the miniboss can only see the player, not hear them.