Crouch-Jump / Crouch-Against wall update
Fixed crouch to jump transition as the player was gaining impulse when jump was pressed while crouching. Jump was disabled if the player is not on walking movement state.
Update Attach to wall functionality as the character was performing a small jump before  attaching to a wall while crouching.
Tested both transitions.