Dash ability iteration
Added a cooldown to the dash.
Added a boolean to set the player immune to damage in the parent player class and configured the GetHir function to take that into consideration.
Set the player to be immune to damage for a short duration after dashing.
Added a damage shield particle system to the player to have a visual representation that the player is immune to damage.