Blue miniboss door AI
Added tasks to toggle the doors when the player enters one of the hallways.
Added a second array of doors to the player so that the right and left doors can be activated separately.
Added blackboard values to determine which side the enemy is teleported to and to toggle the correct doors.