Blue miniboss arena update/Minor AI change
Changed the arena to have a large circle that the miniboss can move along around the arena. The player cannot get to that platform and must interface with the environment to destroy the boss.
Changed the charge time and color in order to better indicate when the enemy is about to fire.
Set the laser sight size back to default when the miniboss loses line of sight on the player.