Blue miniboss platform
Created a static mesh of the platform that the boss moves along.
Added a collider to the mesh.
Created a blueprint that uses that static mesh.
Enabled physics in the blueprint when a condition is met (currently after 3 seconds but it will be once all pillars have explosives planted on them). This allows the platform to fall with the enemy which is then killed by the kill volume in the level which completes the boss fight.