Blue Miniboss pillars update
Created a class for the miniboss pillars with a static mesh for the pillar itself and for an explosive that the player will plant.
Added an input binding and logic to the player class so that they can plant an explosive.
Changed the boss platform blueprint to check if all of the bombs are planted and to explode.
Added an explosion sound effect and particle system for when the pillars explode.