Blue miniboss path mechanic
Created a class for toggle-able actors which are actors with static meshes that can be changed from visible/causes a collision to invisible/does not cause a collision.
Added a function that allows the actor to be toggled on and off.
Created a blueprint from this class and added a static mesh for a wall.
This wall is used to funnel the player into set corridors where they must avoid the boss' shots.
Added an array of walls to the pillars so that they can be toggled when the player plants an explosive.
Created a maze for the player to use until they reach the first pillar.
*Note: this creates large arrays of objects that must be toggled at the same time so it might have an adverse effect on performance and will be monitored*