Blue miniboss path layouts
Added two more layouts for paths to the pillars.
Changed the paths to be one large static mesh instead of many smaller ones.
Changed the pillar to have a "ToggleOn" and "ToggleOff" actor reference so that they could each remove the previous path while turning their own path on.
Configured on and off toggles for each existing path.
Changed the explosive planting text to only trigger if the explosive had not yet been planted.