Blue miniboss mechanic/AI
Added a starting layout for the miniboss arena.
Added two layouts to the miniboss so that he can toggle the starting layout off and layout1 on.
Added a blackboard value to determine that the maze has started and created a new task to toggle the initial layouts and teleport the player to the start of the maze.