Blue miniboss AI
Added the miniboss platform to the miniboss as a variable so that it has access to the array of pillars.
Added booleans to the blackboard in order to keep track of phases.
Changed the AI controller so that the miniboss could set the phase of the fight depending on which pillars had explosives planted on them.
Changed the behavior tree to allow for certain tasks to run, regardless of whether or not the miniboss can see the player.