Blue miniboss AI update
Added an array of targets to the miniboss which act as possible locations that the miniboss can move to.
Added a task that allows the miniboss to move to a target location.
Added a task that allows the miniboss to teleport to a target location.
Stopped the miniboss laser sight from appearing when the player is not detected.
Added a charge shot task for phase 3 which charges faster than the other task. It is currently not in use but may be in the future.