Level setup/Dash fix
Set up a level with a landscape and a base material that might change or might serve as a way to ensure that any floor assets on top of it will not bleed light through small cracks.
Fixed an issue with the dash that was caused by the player being airborne during the dash. The player was not losing speed (as they would be if dashing along the floor) and they were able to go much further than intended with a dash. This was fixed in blueprint by adding a downward impulse on the player if they are in the air.