Asset creation
Created a blueprint to measure the size of assets from the level pack so that I can create other static meshes that will fit properly. The approximate value shown in the editor was not accurate enough.
Created a static mesh of a wall that is 1000 units long and the same width and height as the doorway asset from the pack.