Level assets/Health pickup change
Created a sliding door blueprint using the assets from the asset pack. The door slides open when the player is near and slides back when the player moves away.
Created a short wall static mesh.
Changed the health pickup to not heal the player or become destroyed if the player is at full health.
Added more to the layout of the level.