Ammo pickup/Locked sliding door
Created an ammo pickup which add to the ammo capacity of the player's weapon with a different amount of ammo based on the current weapon equipped.
Created a second sliding door which has a public actor which is the key for the door. The door will only open if their key has been picked up.