Sliding door change/Keycard
Changed the speed in which sliding doors open and close in order to make it quicker.
Created a KeyPickup class which is a pickup that is destroyed on contact with the player and is used as the key for locked doors.
Changed the locked sliding door to use an actor of type KeyPickup.
Imported a model and textures for the keycard and configured its material.