Keycard/Asset creation/Weapon change
The keycard is now larger, facing upward and rotating in place.
Created a room template static mesh so that they can easily be reused when needed.
Change the collision channel for the assault rifle, shotgun (not really needed) and gas mask enemy weapon as they were being stopped by the door triggers.
*Note: I should make a custom channel in order to prevent any future issues with this system but I ran into problems last time I tried that so it will stay as is for now*