Locked sliding door/Door light
Created a new type of sliding door that takes multiple keys. This will replace the previous locked sliding door in all instances.
Created a door light blueprint which has a public door variable and which changes color based on whether or not the door is locked. This was done so that the player can easily determine that a door is locked.