Individual server assets
Took the server asset into blender to split each part individually as only some of the server asset was necessary.
Created two server variants and a server with computer terminal variant.
Set each origin in Blender and exported as .fbx files.
Imported the servers, set up their material and set their build scale to be much smaller (0.016) as the model was massive in Unreal.