Server block assets/Server room objective
Created four types of server blocks, changing the configuration of servers in each block.
Created four types of server blocks which are nearly identical to the previous four but they have a terminal in place of one of the servers.
Created an objective that the player picks up upon entering the server room which tells them to hack the terminals in order to progress.