Server room/Landscape material/Floor textures
Created a layout for the server room with four terminals and a total of 20 server blocks.
Imported some floor textures that can be used with the landscape.
Created a landscape material which uses two of these textures so that I can paint different floor textures in different areas.
Set up the server room with a different floor than the rest of the level.