Interactable actor/Interactable terminal
Created a new class for interactable actors.
The only instance of interactable actor was previously the blue miniboss pillars but now that has changed.
Changed the player class to look for interactable actors rather than the pillars.
Added an FString to the interactable actor class which is the text that should be displayed when the player is nearby.
Changed the HUD to display this text rather than a message for the miniboss pillars.
Created a static mesh for the interactable terminal (a plane) and created an invisible material to add to it.
*Note: I will have to change the miniboss pillar to use the interactable actor class but that will be done when making the boss levels.*