Locked door change/Objective setup
Changed the locked door blueprint to accept any actor as a key rather than only the keycards. This was done so that the terminals can be added as keys for the final door.
Added terminals as keys to the final door so that it will unlock when they are all hacked,
Configured keycards and terminals to be objective targets.