Server room change/Objective change
Moved the location of the server room to be more central in the level.
Added a door to the other side of the server room so that players can enter/exit from both ends (might add an exit to either side).
Added code to the player class function in which they receive an objective so that an objective with the same name will not overwrite their current objective.