Server room spawner/Shotgun change
Added targets that the spawner will spawn enemies on.
-----------------------------------------------------------------------------
*Forgot to commit this and then it changed but I am leaving this so that I can remember my steps for the report.*
Added logic so that enemies are spawned at those targets when the hack progress reaches certain thresholds.
*Note: there is currently a conflict when spawning enemies this way so I'll need to look into it*
-----------------------------------------------------------------------------
Added a group of enemies under the server room which are moved to the spawn locations when needed. This was done to solve the spawning issue that I had and it is probably more performant since the assets are already loaded into the level.
Hid the health bar widgets for each enemy which is under the level and enabled them when they are "spawned". 
The spawned enemies are immediately aware of the player's location and will search for them immediately.
Change the collision channel for the shotgun back to the pawn channel as it was not correctly piercing target.
*Note: I'm not sure why it is not piercing on the other channel but this will work for now. The rerason for changing the channel in the first place was that the assault rifle wouldn't hit enemies when in the sliding door trigger but the shotgun overlaps and hits all actors so it doe not have the same problem*