Blue miniboss arena
Changed the miniboss pillars to work with the new interactable actor as a super class.
Added a collider to the miniboss platform so that it could be destroyed by the kill volume.
Changed the moniboss platform to no longer simulate physics when all explosives have been planted and instead made it update its position while decreasing its z value.
Changed the miniboss to set its dead blackboard value to true when the player plants all explosive so that it does not try to teleport after the platform has begun falling.
Adjusted the player walkable platform slightly as it was not even.