Player starting area/Sliding door fix
Added a room that the player starts in. There are four locked doors which will have health and ammo. Each door will open after killing a miniboss.
Configured the blue and green miniboss as keys to two of the locked doors.
Added miniboss doors which teleport the player to the miniboss matching the color of the door.
Configured the blue and green miniboss doors.
Added the player's starting location as the teleport location for the blue and green minibosses.