Yellow miniboss arena
Created an outline for the yellow miniboss arena.
Changed the static mesh for the shield generators.
Configured the yellow miniboss door.
Linked the yellow miniboss to the generators and vice versa.
Configured the yellow miniboss to be an objective target.
Configured the yellow miniboss to open one of the locked doors in the player's starting room.