Dash fix/Miniboss level lighting/Shield generator change
Fixed the dash as it was improperly removing the laser sight for the duration of its cooldown rather than its period of damage immunity.
Changed many of the existing lights in the miniboss level.
Added lighting to the yellow miniboss arena.
Changed the shield generator's static mesh so that it properly blocks player movement.