Falling rocket change/Red miniboss AI improvement
Changed the color of the visualization for the falling rocket's area to be red rather than white.
Created a new task that spawns falling rockets at the player's location.
Configured the behavior tree to use this node ten times in a loop before and after spawning the falling rockets in the set configurations.
Added another falling rocket to the diagonal rocket spawn function.