Main menu change/improvements
Created a new game mode for the main menu scene.
Created a HUD that uses the widget which was created.
Configured the game mode to use the HUD.
Added a background image to the menu.
Created a level select button.
Changed the color of the buttons on hover.
Imported a sound which is played when hovering over a button.