Main menu level select
Added a back button.
Added each button to an array so that I could easily swap their visibilities.
Created a function that swaps each button's visibility so that they can be turned on/off based on which menu the player is in.
Created placeholder buttons which will take the player to a level based on whichever button they press (currently they are all placeholders with text Level 1 - 6).