Checkpoint prep
Added a bool to check if the player is dead.
Added logic to reload the level on death with an event dispatcher.
Added logic to set the whether or not the player should load the checkpoint when loading a level.
Added logic to set the load checkpoint to false when quitting the game. *Note: this does not seem to be working*