Level checkpoint
Created a new save system that works for a given level.
Added an array of integers to the save game.
Created an array of actors that can be destroyed in the level blueprint.
Set the integers in the save game to 0 if an actor was destroyed.
Set the level blueprint to save the actors who were killed when the enemy dies and to load them at the start of the level if the player died in the level.
Set the array to all 1's (indicating that all actors should be alive) when the player exits from the level.