Main menu hover color/Mutant prep
Changed the main menu hover color from red to orange.
Created a mutant class.
Created a blueprint from that class and configured it.
Created an animation blueprint and blendspace for the mutant.
Created an AI controller, blackboard and behavior tree for the mutant.
Created a task that determines if the mutant is in melee range and damages the player.
Configured the blendspace with some animations and configured it to work with the animation blueprint.