Mutant behavior
Configured the mutant animation blueprint to use an upper body slot.
Trimmed the attack animation.
Created a montage for the attack.
Changed the mutant damage.
Changed the attack task to play the attack montage.
Removed the move node from the behavior tree to the AI controller.
Added logic to stop the mutant's movement and cancel their attack montage when stunned or killed.