Mutant tweaks/Door navigation
Put the mutant back to full speed.
Changed the play rate of the attack montage to speed up the animation.
Changed the mutant so that it takes more damage.
Removed the collision on the mutant when they die.
Added a minimap marker to the mutant.
Updated the door blueprints so that they no longer prevent navigation.