General fixes/End level loading
Added logic to the camera system so that it no longer displays the player's minimap icon.
Removed the mine defusing text.
Added a level loader to load into the correct end level depending on whether or not the player has collected all blueprints.
Added logic to load back to the main menu after a few seconds in either end level.